requirements.txt
Django>=3.1.0
djangorestframework>=3.9.0
flake8>=3.7.0
djangorestframework-jwt==1.11.0
psycopg2>=2.7.5
django-filter==2.4.0